Separando as views contidas em contentview para outros arquivos
